First commit
Service that gets URL as parameter and returns a dictionary containing the IPs of the servers in the route path is done. Service that gets a IP as parameter and returns its location (city, state and country) is done. Integration test with these two services that returns a dictionary containing, for each server, its IP and location is done.